requirements.txt
pytest==5.4.2
python-dotenv==0.12.0
lxml==4.5.0
Flask==1.1.2
requests==2.23.0
konlpy==0.5.2